Unify feeder amount handling behind a FEEDER_COUNT constant
The single-vs-dual hopper split ('a' vs 'a1'/'a2') was hard-coded three
different ways across the feeder Device classes: a per-item key, a
settings key pair, and an `instanceof YumshareDual\Device` check in
DeviceActions.php. Every single-hopper class had its own hand-copied
toFeed()/scaleAmountsForWire(), ported commit-by-commit from D4SH (see
"Port D4SH schedule fixes to ..." in the log).

Extract that logic into HandlesFeederSchedule, driven by a
`FEEDER_COUNT` constant (default 1 on DeviceDefinition, 2 on
YumshareDual - schedule.md §4e/§4f). Fixes two real drifts that came
from D3 (FreshElement3) never receiving the manual port: it used
last($latest) instead of head() for nextTick (farthest upcoming feed
instead of nearest) and was missing the +1 offset. Also gives D3/D4
a working toFeedGet(), which only D4H/D4SH had before even though
DevFeedGetController calls it unconditionally on any device type.
