+++ app/Petkit/DeviceDefinition.php
@@ -7,6 +7,17 @@ use App\Models\Device;
 
 interface DeviceDefinition
 {
+    /**
+     * How many hopper amounts this device's feed schedule carries per item -
+     * 1 (wire key 'a') for every feeder except D4SH, which is 2 ('a1'/'a2').
+     * schedule.md §4e/§4f is the single source of truth for this split;
+     * HandlesFeederSchedule reads this constant instead of any class
+     * hard-coding 'a1'/'a2' or an `instanceof` check on a specific device.
+     * Non-feeder devices (litter boxes, fountains) don't use this constant
+     * but inherit the default so it's never undefined.
+     */
+    public const FEEDER_COUNT = 1;
+
     public function hasAction(string $action): bool;
     public function stateTopics(): array;
     public function subscribedTopics(): array;
